#6 Move logging to own file
